refactor (web): move shuffle array method to colour store
