+++ apps/web/src/stores/colourStore.ts
@@ -1,6 +1,13 @@
 import { INITIAL_COLOURS } from "@/utils/constants";
-import { shuffleArray } from "@/utils/helpers";
 import { create } from "zustand";
+
+export const shuffleArray = (array: string[]) => {
+  for (let i = array.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [array[i], array[j]] = [array[j] as string, array[i] as string];
+  }
+  return array;
+};
 
 interface ColoursState {
   colours: string[];
